Flyttet kode til egen service
